# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.MonitorResponse:
    properties:
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      interval_seconds:
        type: integer
      last_checked_at:
        type: string
      name:
        type: string
      target:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  dto.RegisterMonitorRequest:
    properties:
      address:
        description: 도메인 or IP
        type: string
      interval_seconds:
        minimum: 10
        type: integer
      name:
        type: string
      port:
        description: 포트 번호
        type: string
      type:
        enum:
        - http
        - https
        - websocket
        - tcp
        type: string
    required:
    - address
    - interval_seconds
    - name
    - port
    - type
    type: object
  dto.ResponseFormat:
    properties:
      data: {}
      error_code:
        type: integer
      error_message:
        type: string
      message:
        type: string
    type: object
...
  dto.UpdateMonitorRequest:
    properties:
      address:
        type: string
      interval_seconds:
        type: integer
      name:
        type: string
      port:
        type: string
      type:
        type: string
    type: object
...
host: 10.30.8.25:8888
info:
  contact: {}
  description: keeplo
  title: keeplo API
  version: "0.1"
paths:
...
  /monitor:
    get:
      consumes:
      - application/json
      description: 사용자가 등록한 모든 모니터링 항목을 조회합니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ResponseFormat'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.MonitorResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
      summary: 모니터링 목록 조회
      tags:
      - monitor
    post:
      consumes:
      - application/json
      description: 모니터링 항목을 신규 등록합니다.
      parameters:
      - description: 신규 모니터링 등록 요청
        in: body
        name: monitor
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterMonitorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
      summary: 모니터링 추가
      tags:
      - monitor
  /monitor/{id}:
    delete:
      consumes:
      - application/json
      description: 특정 모니터링 항목을 삭제합니다.
      parameters:
      - description: 모니터링 고유 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
      summary: 모니터링 삭제
      tags:
      - monitor
    get:
      consumes:
      - application/json
      description: 특정 모니터링 항목의 상세 정보를 조회합니다.
      parameters:
      - description: 모니터링 고유 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ResponseFormat'
            - properties:
                data:
                  $ref: '#/definitions/dto.MonitorResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
      summary: 모니터링 상세 조회
      tags:
      - monitor
    put:
      consumes:
      - application/json
      description: 기존 모니터링 항목의 정보를 수정합니다.
      parameters:
      - description: 모니터링 고유 ID
        in: path
        name: id
        required: true
        type: string
      - description: 수정할 정보
        in: body
        name: monitor
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMonitorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
      summary: 모니터링 수정
      tags:
      - monitor
  /monitor/{id}/toggle:
    patch:
      consumes:
      - application/json
      description: 모니터링 항목을 활성화 또는 비활성화합니다.
      parameters:
      - description: 모니터 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
      summary: 모니터링 ON/OFF 전환
      tags:
      - monitor
  /monitor/{id}/trigger:
    post:
      description: 선택한 모니터링 항목을 즉시 테스트 실행합니다.
      parameters:
      - description: 모니터 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ResponseFormat'
      summary: 모니터링 수동 실행
      tags:
      - monitor
...
schemes:
- http
swagger: "2.0"